Add hash map op bench
