--- benchmark/hash_map/.gitignore
@@ -1,10 +1,0 @@
-.vscode
-package.json
-*.o
-*.skel.json
-*.skel.yaml
-package.yaml
-ecli
-.output
-test
-uprobe
