--- benchmark/hash_map/uprobe.c
@@ -1,68 +1,0 @@
-// SPDX-License-Identifier: (LGPL-2.1 OR BSD-2-Clause)
-/* Copyright (c) 2020 Facebook */
-#include <signal.h>
-#include <stdio.h>
-#include <time.h>
-#include <stdint.h>
-#include <sys/resource.h>
-#include <bpf/libbpf.h>
-#include <bpf/bpf.h>
-#include <unistd.h>
-#include <stdlib.h>
-#include "uprobe.skel.h"
-#define warn(...) fprintf(stderr, __VA_ARGS__)
-
-static int libbpf_print_fn(enum libbpf_print_level level, const char *format,
-			   va_list args)
-{
-	return vfprintf(stderr, format, args);
-}
-
-static volatile bool exiting = false;
-
-static void sig_handler(int sig)
-{
-	exiting = true;
-}
-
-int main(int argc, char **argv)
-{
-	struct uprobe_bpf *skel;
-	int err;
-
-	/* Set up libbpf errors and debug info callback */
-	libbpf_set_print(libbpf_print_fn);
-
-	/* Cleaner handling of Ctrl-C */
-	signal(SIGINT, sig_handler);
-	signal(SIGTERM, sig_handler);
-
-	/* Load and verify BPF application */
-	skel = uprobe_bpf__open();
-	if (!skel) {
-		fprintf(stderr, "Failed to open and load BPF skeleton\n");
-		return 1;
-	}
-
-	/* Load & verify BPF programs */
-	err = uprobe_bpf__load(skel);
-	if (err) {
-		fprintf(stderr, "Failed to load and verify BPF skeleton\n");
-		goto cleanup;
-	}
-	err = uprobe_bpf__attach(skel);
-	if (err) {
-		fprintf(stderr, "Failed to attach BPF skeleton\n");
-		goto cleanup;
-	}
-
-	while (!exiting) {
-		sleep(1);
-		printf("loaded ebpf program...\n");
-	}
-cleanup:
-	/* Clean up */
-	uprobe_bpf__destroy(skel);
-
-	return err < 0 ? -err : 0;
-}
